dom/html/test/forms/test_interactive_content_in_label.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=229925
-->
<head>
  <title>Test for Bug 229925</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=229925">Mozilla Bug 229925</a>
<p id="display"></p>
<form action="#">
  <label>
    <span id="text">label</span>
    <input type="button" id="target" value="target">

    <a id="yes1" href="#">a</a>
    <audio id="yes2" controls></audio>
    <button id="yes3">button</button>
    <details id="yes4">details</details>
    <embed id="yes5">embed</embed>
    <iframe id="yes6" src="data:text/plain," style="width: 16px; height: 16px;"></iframe>
    <img id="yes7" src="data:image/png," usemap="#map">
    <input id="yes8" type="text" size="4">
    <keygen id="yes9">
    <label id="yes10">label</label>
    <object id="yes11" usemap="#map">object</object>
    <select id="yes12"><option>select</option></select>
    <textarea id="yes13" cols="1" rows="1"></textarea>
    <video id="yes14" controls></video>

    <audio id="no1"></audio>
    <img id="no2" src="data:image/png,">
    <input id="no3" type="hidden">
    <object id="no4">object</object>
    <video id="no5"></video>
    <span id="no6" tabindex="1">tabindex</span>
  </label>
</form>
<script class="testbody" type="text/javascript">

/** Test for Bug 229925 **/

var target = document.getElementById("target");

var yes_nodes = [
  document.getElementById("yes1"),
  document.getElementById("yes2"),
  document.getElementById("yes3"),
  document.getElementById("yes4"),
  document.getElementById("yes5"),
  document.getElementById("yes6"),
  document.getElementById("yes7"),
  document.getElementById("yes8"),
  document.getElementById("yes9"),
  document.getElementById("yes10"),
  document.getElementById("yes11"),
  document.getElementById("yes12"),
  document.getElementById("yes13"),
  document.getElementById("yes14"),
];

var no_nodes = [
  document.getElementById("text"),
  document.getElementById("no1"),
  document.getElementById("no2"),
  document.getElementById("no3"),
  document.getElementById("no4"),
  document.getElementById("no5"),
  document.getElementById("no6"),
];

var target_clicked = false;
target.addEventListener("click", function() {
  target_clicked = true;
});

var node;
for (node of yes_nodes) {
  target_clicked = false;
  node.click();
  is(target_clicked, false, "mouse click on interactive content " + node.nodeName + " shouldn't dispatch event to label target");
}

for (node of no_nodes) {
  target_clicked = false;
  node.click();
  is(target_clicked, true, "mouse click on non interactive content " + node.nodeName + " should dispatch event to label target");
}

</script>
</pre>
</body>
</html>
